Refactor : generate service class
